fix resizing of root filesystem
resizing of /root was failing to occur.
Previously, there should have been output of resize2fs in 
/tmp/resize.out but all we were seeing there was the initial resize2fs
version info.

I believe the backgrounded process was getting cleaned up somehow
after rc.sysinit exited.  Here we're using start-stop-daemon to
be more explicit that we want this backgrounded.

In the future we could make this operation be configurable (on, off, or
foreground) by having /etc/init.d/resizefs read /etc/default/resizefs
or some other config file.

Then, that config file could be provided through the 'injected files'
in openstack or be written in some other way.

Also added here is 'time_call' which uses uptime to time an operation
to hundredths of a second.  We use that for a nice summary.